Update install-umbraco-with-nuget.md
UmbracoCms.targets includes everything under App_Plugins for instance.
We can't throw Umbraco Forms data around between publishes, so we need to exclude that folder.
I'm sure others will have the same need, so it might as well be in the docs. :)
Took me a while to figure out the culprit at least.
Dunno if this is the best page, but I think it's only relevant for nuget users.