Update incorrect instructions.
ImageProcessor.Web does not require the VirtualPathProvider to be disabled for the cache to work. Disabling the provider is strongly discouraged as it will cause Umbraco to persist the urls in the database in absolute format. This means that ImageProcessor.Web cannot intercept the query without using the /remote.axd prefix.

ImageProcessor.Web.Plugins.AzureBlobCache will create the cache folder if it doesn't already exist.

https://github.com/JimBobSquarePants/ImageProcessor/blob/V2/src/Plugins/ImageProcessor.Web/ImageProcessor.Web.Plugins.AzureBlobCache/AzureBlobCache.cs#L100