Update IContentFinder.md
Improve the example implementation of the IContentFinder by retrieving the 404 page and root page (homepage) with respect to the current culture. If a 404 page was not found then the fallback would be to redirect to the homepage. As last resort is to get all content at root and redirect to a page that has a template