updates PR#545 with up to date files
See #545

Fixes headers for github's markdown parser. Any string in the repo
starting with a hashtag but not followed by either whitespace or hash
character is updated so that they've got a space following them. Should
ignore hashtags in angular URLs, should ignore any hashtag where there's
already a space or another hashtag following it

regex to find was ^([#]{1,})(?!#)(?!\s)(.*)

regex to replace was ${1} ${2}