Check framework version before ensuring MVC
So when creating the project the check for the version of the .NET framework should be done before selecting the MVC project. When I did it I first choose the project type, the .net framework version and then on the second screen I can choose "MVC". A minor change that is probably less confusing for those who read the documentation to follow.