#170 documentation for working with tags (#900)
* Reuse a bit of the repo instruction from core

Untill now I have not needed to clone the repository in order to create PR's for the documentation. But now that I need to do a little bigger contribution (adding a new page etc.) I thought it would be better to clone the repo. Then I realized nothing is currrently mentioned about how to keep the fork in sync with the main repository like it is here on the Core repo: https://github.com/umbraco/Umbraco-CMS/edit/dev-v7/docs/CONTRIBUTING_DETAILED.md - So thought that I would add some details about it here too heavily inspired by the core text.

* WIP: Initial commit for tags documentation

* Adding some text descriptions - Needs image examples and code samples + some minor adjustments/proof reading

* Finishing up tags documentation for 7.6+

* Make sure alias and return type is on seperate lines

* WIP: Add markdown file for legacy tags documentation - Need to edit what is already there though

* Fix naming convention for legacy tags documentation

* Remove v7.0.0 stuff, which will be in a seperate PR

* Add deocumentation for the Tag group setting

* Remove v7 images since the will be featured in another PR

* Fixed internal page reference