Updated RelationType explanations
Added meta data. 

Added more explanation on the classes purpose and corrected my use of Id to Key.

Do I need to add the properties from the base `Entity` class? As these are used in a lot of places would it be better creating a new page and adding references back to it.