Attempt to update the code examples using the appropiate syntax
Based on the preview I can't figure out if the code examples that are a part of an indented list are actually going to be parsed and rendered correctly? Or it this something that needs to be fixed somewhere? - Anyway I have tried to markup the code examples using the appropiate syntax 😃