Update v8documentation.md
We need to be consistent with the naming of all Umbraco 8 related communication - including documentation. That's why it should always be Umbraco 8 not v8. The title of this doc then also needs to be changed and the same goes with all the urls and possibly also the copy in those...