Markup code samples as XML (#1361)
* Markup code samples as XML

I'm wondering if perhaps this needs to be versioned? I don't know if the XSLT example works in Umbraco 7.13 these days? But anyhow we might as well prepare for the release of v8 where this way will no longer be supported...right? 😄

* Add YAML annotation about versions

So it turns out my memory was wrong - The XsltXtensions.config was removed with the release of Umbraco 7.0.0 😃 So I have added the YAML annotation with the versioning from 4.5 to 6.2.6 - I made sure to check some old builds and Examine was introduced in the v4.5 release.

Also marked up the other examples using the XML syntax since they're parts of some XML even though not the full code is displayed - Probably for simplicity so it's fine 😃

* Fix XSLT casing

* Make a vpre7 on the old xslt-example and add a new v7 example. This example is not ready yet though since it needs a more proper descripton and perhaps a more "full" example of the XSLT file. Fix author commit name.

* Revert "Make a vpre7 on the old xslt-example and add a new v7 example. This example is not ready yet though since it needs a more proper descripton and perhaps a more "full" example of the XSLT file. Fix author commit name."

This reverts commit fd88a0eb07396d379695d86c5bf312aefcaf5262.

* Add a version pre v7 and a new version about v7

* Revert changes

* Add a pre version 7 file and ad "current" v7 example removing steps 1 to 3. Still needs some work.

* Temporarily set from and to version to just 7

* update version removed