Update index.md
On inspecting the RedirectTrackingComponent component, which is the v8 equivalent of RedirectTrackingEventHandler I've noticed there no longer is a check for the existence of a 3rd party DLL with redirect functionality in the solution. Changed the documentation accordingly, I don't know of this is by design or that the check is simply missing in the v8 codebase.