Updated surface-controllers.md
Added information regarding where to put controllers in a plugin based controller. It was not obvious that it cannot go in the App_plugins folder too, which took me a few days to figure out why.

If there is a link to a page in the documentation where this is detailed, I will be happy to add that too, but I cannot find it.