Issues upgrading from a version prior to 7.6
If you upgrade your project from a version prior to 7.6 without updating your property editors to the non-obsolete version, that stores UDIs rather than IDs you might run into issues when trying to transfer content to Umbraco Cloud as these pickers will have the wrong content. It is therefore important that customers know that if they want a project prior to v7.6 on Cloud, they might need to do a db migration as well - if they're using the obsolete pickers.