Update Control-Generation.md with more info on installing full versions of ModelsBuilder
I struggled a bit this week after reading this to understand the best way to install the full version of models builder as nuget doesn't seem to be referenced in the ModelsBuilder repository, and it took me a while to figure out which nuget packages to install.

So I thought I'd submit a PR with more detailed instructions in case anyone else has the same problem!  I also wanted to include a link to the package on Our, but I couldn't find it?  Only the old version of ModelsBuilder?