Added detail to how different "Lifetimes" work
I recently used `Lifetime.Request`, assuming that this is would create a new instance per web-request.  It turned out I actually should have used `Lifetime.Scope`.

This is because the terminology is based on Lightinject Lightinject (which uses "Request" and "Scope" lifetimes), which is confusing in the context of a web application - this was highlighted and clarified by @Shazwazza directly with Lightinject this this issue he raised: https://github.com/seesharper/LightInject/issues/494
It has also been raised and answered in the forums (e.g here: https://our.umbraco.com/forum/umbraco-8/98389-iusercomposer-with-lifetimerequest-works-as-lifetimetransient)

This PR is to update the IoC docs to add more information around the differences between the different lifetimes for those unfamiliar with the nuances in the terminology.