Fixed a few typos and some sentence structure.
Line 186 - However, once you've checked this permission setting on one Document Type, it will be the **only one** that can be used to create content at the root.

That is not correct as far, as I know. You can have several different doc types that can be used as root content. Please, remove that line unless I somehow misunderstood what it means @sofietoft