Unattended-Install CLI arguments are incorrect
The arguments for the CLI example don't work. the arguments are case sensitive and should be lower case. Also the `\` in the connection string should only be single as it gets escaped when added to the AppSettings Connection string. Entering `\\` caused the connection string to have 4 slashes `\\\\`  

You can verify the arguments against the (Install page)[] which are correct - https://our.umbraco.com/documentation/Fundamentals/Setup/Install/install-umbraco-with-templates