Verify article for v10
Adding a couple of `@using` statements that was missing from the C# sample - this also applies to the latest version of 9.

There is one difference between v10 and v9 when testing this article.
For v10, VS suggests adding a `?` to the main method in the C# sample:

```
public ContentApp? GetContentAppFor(object source, IEnumerable<IReadOnlyUserGroup> userGroups)
        {
```

It seems to be related to the nullable references.
When adding the same to the v9 code, it still works, only gives a warning.