update activity log function
By following and to some extend understanding the instructions on this brilliant tutorial I managed to create a friendly welcome screen for my editors. The dashboard includes the suggested activity log which is a nice feature.

However, if editors save or publish a page several times, it will appear on the "pick up where you left off" list several times. Ideally, a node would only appear on the list once, regardless of multiple saves or publishes.

The activity log function seems to attempt to take care of this, but for some reason it does not, at least not in my case. 

So being a Angular JS rookie I tried to rewrite the loop to better understand what is going on. And this code actually works. Maybe other users could benefit from this.