Update wording of explanation of default Content Version Cleanup
Had a question on the forum asking what happens after 90 days as existing wording sort of hinted the document would be deleted...
I thought restructuring it like this might make it clearer that Published content is never deleted? The settings are a bit of a mouthful, so probably only so much we can do!