Example of how to create async RenderController
It was hard to find any documented way to use an async Index()-method but this approach works. I think that official docs should have an example for people looking to use a async implementation in the controller. 

Without the sync Index() with the [NoAction]-attribute the call would be ambiguous between the async version on the ProductPageController and the sync Index()-method on RenderController.

As far as I can tell this is the best approach at the moment.