Add section on runtime minification
We had deployed an update to our site that included custom macro parameter editors, but the editor wasn't loading the AngularJS controller at runtime. After much head scratching, we found this post: https://our.umbraco.com/forum/umbraco-9/107740-umbraco-9-cant-register-controller-for-a-custom-dashboard-that-works-locally-but-fails-when-deploy
which described the issue and the solution. I figured it would be valuable to include a note about this on this documentation page for others.