Update rewrite rules docs
We need to enrich the www redirect rule to contain ".*azurewebsites.net*" pattern, otherwise we will have issues deploying as the rule will redirect to www.{stie}.azurewebsites.net which will break the deployment process.

Fixing this in the site extension is a no-go, as we don't want to hijack the customer's rewrite rules.

DISCUSSION:
Do we want to include a warning hint mentioning this? 🤔 
I've added it but can be removed or updated if it does not make sense.