removed Visual Studio
we have had cases where clients try to pull and push changes to and from Cloud and confirmed it is slower than it should be and are asking us to help with this, which is not something we can do. We always recommend clients to use a git client such as Fork, SourceTree, Gitkraken etc as they are faster and work as expected.