Fix Custom Search Indexing Sample 2
The sample on this page won't work for creating a filtered index by 'product', this is because the core 'ContentValueSetValidator' is used  in the example to apply the 'validator' rules to the index options. and this validator requires there to be both a 'path' and a value for '__Published' in the index to apply it's validation.

https://github.com/umbraco/Umbraco-CMS/blob/0765a2ebbb6fd452b6543a349f5d0c1f157e5588/src/Umbraco.Infrastructure/Examine/ContentValueSetValidator.cs#L155

But the example has removed those fields from the index. by defining a custom list of FieldDefinitions and a custom populator.

This PR adds those two fields into this custom index, which is the minimum required for the example to work 'as is'.