Update two-factor-authentication.md
I noticed that the namespace was Umbraco9, so I changed it to something more generic and as per https://github.com/umbraco/UmbracoDocs/blob/main/12/umbraco-cms/implementation/composing.md.

I don't know if there is a best practice for what namespace to use? But I thought it best not to be version specific, anyway.