Change information about using CNAME records pointing to dns.umbraco.io
Wildcard asterisks in DNS records, are no longer supported when combined with dns.umbraco.io as that will make SSL certificate generation and renewal fail. 

See support ticket #72019 

This change also clarifies that A & AAAA records has to be used for root domain, as CNAME records are not possible for root domains.