Update ui-config-files.md
If you are hosting your site in Umbraco Cloud and using Visual Studio then you will have to make sure that the file/files (or parent folder of the files) you include will be configured in certain way in the .Web project configuration file. This is to ensure that the Umbraco Cloud site will be able to use the file/files.