Specifying that Umbraco includes NPoco
This documentation worked really well for me, however I wished that it had specified that we're using NPoco. In earlier Umbraco days, it used to be PetaPoco, of which NPoco is based on.

I think it's also important to mention it explicitly and link to it, as developers are likely going to need to go to NPoco's own documentation to learn about it's own api.